Extracted a hook from ShortURLList component
